Fixed time display on index and show!

+++ db/migrate/20150328205002_create_recipes.rb
@@ -3,7 +3,7 @@ class CreateRecipes < ActiveRecord::Migration
     create_table :recipes do |t|
       t.string :name
       t.text :directions
-      t.float :cooking_time
+      t.integer :cooking_time
 
       t.timestamps null: false
     end
